Move the util class
